+++ inyange-dashboard/src/app/(dashboard)/dashboard/components/Barchart/barchart.tsx
