+++ inyange-dashboard/src/app/(dashboard)/dashboard/hooks/uselogin.ts
@@ -3,17 +3,17 @@ import { Credentials } from "../utils/types";
 
 export const useLogin = () => {
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const [errorMessage, setErrorMessage] = useState(""); // Initialize error message state
-  const [successMessage, setSuccessMessage] = useState(""); // Initialize success message state
+  const [errorMessage, setErrorMessage] = useState(""); 
+  const [successMessage, setSuccessMessage] = useState(""); 
 
   const login = async (credentials: Credentials) => {
     setIsSubmitting(true);
-    setErrorMessage(""); // Reset previous error message
-    setSuccessMessage(""); // Reset previous success message
+    setErrorMessage("");
+    setSuccessMessage(""); 
 
     try {
       const response = await fetch(
-        "https://buildmart-42eabdb55b17.herokuapp.com/api/login/", // Adjust this URL based on your environment
+        "https://buildmart-42eabdb55b17.herokuapp.com/api/login/", 
         {
           method: "POST",
           body: JSON.stringify(credentials),
@@ -24,28 +24,28 @@ export const useLogin = () => {
       );
 
       if (!response.ok) {
-        const errorResponse = await response.json(); // Parse as JSON
+        const errorResponse = await response.json();
         setErrorMessage(
           `Login failed: ${errorResponse.message || "Unknown error"}`
-        ); // Set a better error message
+        ); 
         throw new Error(
           `Login failed: ${errorResponse.message || "Unknown error"}`
         );
       }
 
       const data = await response.json();
-      setSuccessMessage("Login successful!"); // Set success message on successful login
-      return data; // Return data on successful login
+      setSuccessMessage("Login successful!"); 
+      return data;
     } catch (error) {
       console.error("Failed to login:", error);
       setErrorMessage(
         (error as Error).message || "An error occurred during login."
-      ); // Handle the error
-      throw error; // Propagate the error to be handled by the caller
+      );
+      throw error; 
     } finally {
-      setIsSubmitting(false); // Reset submitting state
+      setIsSubmitting(false); 
     }
   };
 
-  return { login, isSubmitting, errorMessage, successMessage }; // Return the necessary properties
+  return { login, isSubmitting, errorMessage, successMessage }; 
 };
