+++ inyange-dashboard/src/app/api/orderdetails/route.ts
@@ -5,9 +5,13 @@ const baseURL = process.env.BASE_URL;
 
 export async function GET() {
   try {
-    const response = await fetch(`${baseURL}/api/orderdetails`);
+    // const response = await fetch(`${baseURL}/api/orderdetails`);
 
-    
+    const response = await fetch(`${baseURL}/api/orderdetails/`, {
+      next: { revalidate: 10 }, 
+    });
+
+
     if (!response.ok) {
       const errorText = await response.text(); 
       return NextResponse.json({ error: 'Failed to fetch order details: ' + errorText }, { status: response.status });
