+++ inyange-dashboard/src/app/page.tsx
@@ -1,10 +1,9 @@
-import SignUpForm from "./register/page";
+import SignUpForm from "./(dashboard)/dashboard/register/page";
 
 export default function Home() {
   return (
-<div>
-<SignUpForm /> 
-
-</div>
-  )
+    <div>
+      <SignUpForm />
+    </div>
+  );
 }
